function that prints all the elements of a list_t list.

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
 
 size_t print_list(const list_t *h)
 {
-	unsigned int i = 0;
+	size_t i = 0;
 
 	while (h != NULL)
 	{
@@ -16,7 +16,7 @@ size_t print_list(const list_t *h)
 			printf("[0] (nill)");
 		else
 		{
-			printf("[%u] %s\n", h->len, h->str);
+			printf("[%d] %s\n", h->len, h->str);
 		}
 		i++;
 		h = h->next;
